removed original RadialTree Component and replaced it with PsqlGraph component
